Upload fSIR results as HTML
Functional Sliced Inverse Regression is a dimension reduction technique that separates the training data into H slices, grouping subjects by their output, in this case, Systolic Blood Pressure. Then, the inverse regression of the functional data given Systolic Blood Pressure is observed to create and train the model. The model produces several sufficient predictors that simplify the functional data, with decreasing proportions of relationship. We found that the model performed best using two of these sufficient predictors, and H = 5 slices. Like with all models, we used 10-fold cross-validation, where the data was split into ten folds, then the model would train on 9 of these folds, and test on the other. This process is repeated so that each fold tests once. We also smoothed the functional data using 20 basis functions, consistent with our other models. To measure performance, we trained a linear regression model using the sufficient predictors and scalar variables.