Functional Random Forest
HTML: https://rpubs.com/darrentsumm/NHANESFRF

Random Forest is a machine learning model that is centered around decision trees. A decision tree starts with a binary statement (e.g., “Is the patient’s Age greater than 45?”) and two leaves are created, one for yes, one for no. This branching continues until the tree reaches a terminal node, which assigns a prediction, either categorical or continuous.
A random forest contains a set number of decision trees; for our model, 500.  These trees are created randomly by the model, based on the variable we set, mtry, which we often set as 5. This means that each tree will randomly select 5 variables without replacement to create a tree, and use a bootstrap sample of the training data to find the best splitting point, i.e., the number after “greater than”. For RFs with continuous outputs, like Systolic Blood Pressure, the outputs of all 500 trees are averaged together for each subject. 
Since we modified the functional data using bsplines, each of the 20 basis points counted as a variable in this model.